pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "setuptools-scm>=8.0"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]  # exclude packages matching these glob patterns
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

[tool.setuptools.package-data]
nodenorm = ["*.json"]

[project]
name = "nodenormalization"
version = "0.0.1"
authors = [
  { name="Johnathan Schaff", email="jschaff@scripps.edu" },
]
description = "Translator Node-Normalization Service"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Utilities",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "biothings[web_extra]==1.0.2",
    "GitPython>=3.1.43",
    "bmt>=1.4.5",
]


[project.optional-dependencies]
tests = [
    "pytest >= 8.1.1",
    "pytest-asyncio >= 0.23.8",
    "pytest-randomly",
]

telemetry = [
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "opentelemetry.exporter.jaeger.thrift>=1.21.0",
    "opentelemetry-instrumentation>=0.42b0",
    "opentelemetry-instrumentation-tornado>=0.42b0",
    "deprecated>=1.2.0"
]


[project.urls]
"Homepage" = "https://github.com/biothings/biothings_annotator"


[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311', 'py312', 'py313', 'py314']
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
)
'''


[tool.isort]
profile = "black"
combine_as_imports = true
line_length = 120
src_paths = ["."]
extend_skip = []



# pytest configuration
[tool.pytest.ini_options]
minversion = "6.2.5"
pythonpath = ["."]


# Options
addopts = [
    "-rA",
    "-vv",
    "--doctest-modules",
    "--setup-show",
    "--capture=no",
    "--tb=line",
    "--durations=0",
    "--showlocals",
    "--strict-markers",
    "--color=yes",
    "--code-highlight=yes"
]


# Path
norecursedirs = [
    ".svn",
    ".git",
    "_build",
    "tmp*",
    "lib",
    "lib64",
]
testpaths = ["test"]


# Markers
markers = [
    "unit",
    "performance"
]


# Logging
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"


# Asyncio
asyncio_mode = "auto"
